p/Add time spent indicators and legend to Work Planned Calendar Grid
- Introduced TimeSpentIndicator and TimeSpentLegend components for visualizing time spent against estimates.
- Updated WorkPlannedCalendarGrid to include actual start and end dates.
- Enhanced WorkPlannedEntry interface to accommodate actual start and end date fields.
- Added localization for actual start and end date labels.
