+++ apps/gotask_ui/src/i18n/messages/en.json
@@ -1008,6 +1008,8 @@
     "leave": "LEAVE",
     "permission": "PERMISSION",
     "actualtime": "Actual Time Spent",
+    "actualstartdate": "Actual Start Date",
+    "actualenddate": "Actual End Date",
     "meta": {
       "title": "Work Planned | GoTaskToday",
       "description": "Schedule and monitor planned work activities and milestones."
